Backup: Pi values nicely corelate to correct action
If the nn works reasonably well then the Pi values are looking
pretty good. The replacement of NN with MCTS allowed to fix issues and
tune the MCTSNN as well as the NN as well.

Q: With MCTS backend, if training data is generated and then used to
train NN. Does that set a good starting benchmark for the NN?